feat(qtum): introduce qtum_delegation module
