+++ mm2src/coins/qrc20.rs
@@ -1,7 +1,7 @@
 use crate::eth::{self, u256_to_big_decimal, wei_from_big_decimal, TryToAddress};
 use crate::qrc20::rpc_clients::{LogEntry, Qrc20ElectrumOps, Qrc20NativeOps, Qrc20RpcOps, TopicFilter, TxReceipt,
                                 ViewContractCallType};
-use crate::utxo::qtum::{QtumBasedCoin, QtumCoin};
+use crate::utxo::qtum::QtumBasedCoin;
 use crate::utxo::rpc_clients::{ElectrumClient, NativeClient, UnspentInfo, UtxoRpcClientEnum, UtxoRpcClientOps,
                                UtxoRpcError, UtxoRpcFut, UtxoRpcResult};
 use crate::utxo::utxo_common::{self, big_decimal_from_sat, check_all_inputs_signed_by_pub, UtxoTxBuilder};
@@ -9,11 +9,11 @@ use crate::utxo::{qtum, sign_tx, ActualTxFee, AdditionalTxData, BroadcastTxErr, 
                   HistoryUtxoTx, HistoryUtxoTxMap, RecentlySpentOutPoints, UtxoActivationParams, UtxoCoinBuilder,
                   UtxoCoinFields, UtxoCommonOps, UtxoFromLegacyReqErr, UtxoTx, UtxoTxBroadcastOps,
                   UtxoTxGenerationOps, VerboseTransactionFrom, UTXO_LOCK};
-use crate::{BalanceError, BalanceFut, CoinBalance, DelegationError, DelegationResult, FeeApproxStage,
-            FoundSwapTxSpend, HistorySyncState, MarketCoinOps, MmCoin, NegotiateSwapContractAddrErr, SwapOps,
-            TradeFee, TradePreimageError, TradePreimageFut, TradePreimageResult, TradePreimageValue,
-            TransactionDetails, TransactionEnum, TransactionFut, TransactionType, ValidateAddressResult,
-            WithdrawError, WithdrawFee, WithdrawFut, WithdrawRequest, WithdrawResult};
+use crate::{BalanceError, BalanceFut, CoinBalance, DelegationError, FeeApproxStage, FoundSwapTxSpend,
+            HistorySyncState, MarketCoinOps, MmCoin, NegotiateSwapContractAddrErr, SwapOps, TradeFee,
+            TradePreimageError, TradePreimageFut, TradePreimageResult, TradePreimageValue, TransactionDetails,
+            TransactionEnum, TransactionFut, TransactionType, ValidateAddressResult, WithdrawError, WithdrawFee,
+            WithdrawFut, WithdrawRequest, WithdrawResult};
 use async_trait::async_trait;
 use bigdecimal::BigDecimal;
 use bitcrypto::{dhash160, sha256};
@@ -27,7 +27,7 @@ use common::mm_error::prelude::*;
 use common::mm_number::MmNumber;
 use common::now_ms;
 use derive_more::Display;
-use ethabi::{Contract, Function, Token};
+use ethabi::{Function, Token};
 use ethereum_types::{H160, U256};
 use futures::compat::Future01CompatExt;
 use futures::lock::MutexGuard as AsyncMutexGuard;
@@ -57,25 +57,15 @@ mod swap;
 pub const OUTPUT_QTUM_AMOUNT: u64 = 0;
 pub const QRC20_GAS_LIMIT_DEFAULT: u64 = 100_000;
 const QRC20_PAYMENT_GAS_LIMIT: u64 = 200_000;
-pub const QRC20_GAS_LIMIT_DELEGATION: u64 = 2_250_000;
 pub const QRC20_GAS_PRICE_DEFAULT: u64 = 40;
-const QRC20_DUST: u64 = 0;
+pub const QRC20_DUST: u64 = 0;
 // Keccak-256 hash of `Transfer` event
 const QRC20_TRANSFER_TOPIC: &str = "ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef";
 const QRC20_PAYMENT_SENT_TOPIC: &str = "ccc9c05183599bd3135da606eaaf535daffe256e9de33c048014cffcccd4ad57";
 const QRC20_RECEIVER_SPENT_TOPIC: &str = "36c177bcb01c6d568244f05261e2946c8c977fa50822f3fa098c470770ee1f3e";
 const QRC20_SENDER_REFUNDED_TOPIC: &str = "1797d500133f8e427eb9da9523aa4a25cb40f50ebc7dbda3c7c81778973f35ba";
-pub const QTUM_ADD_DELEGATION_TOPIC: &str = "a23803f3b2b56e71f2921c22b23c32ef596a439dbe03f7250e6b58a30eb910b5";
-pub const QTUM_REMOVE_DELEGATION_TOPIC: &str = "7fe28d2d0b16cf95b5ea93f4305f89133b3892543e616381a1336fc1e7a01fa0";
-const QTUM_DELEGATE_CONTRACT_ABI: &str = r#"[{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"_staker","type":"address"},{"indexed":true,"internalType":"address","name":"_delegate","type":"address"},{"indexed":false,"internalType":"uint8","name":"fee","type":"uint8"},{"indexed":false,"internalType":"uint256","name":"blockHeight","type":"uint256"},{"indexed":false,"internalType":"bytes","name":"PoD","type":"bytes"}],"name":"AddDelegation","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"_staker","type":"address"},{"indexed":true,"internalType":"address","name":"_delegate","type":"address"}],"name":"RemoveDelegation","type":"event"},{"constant":false,"inputs":[{"internalType":"address","name":"_staker","type":"address"},{"internalType":"uint8","name":"_fee","type":"uint8"},{"internalType":"bytes","name":"_PoD","type":"bytes"}],"name":"addDelegation","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"delegations","outputs":[{"internalType":"address","name":"staker","type":"address"},{"internalType":"uint8","name":"fee","type":"uint8"},{"internalType":"uint256","name":"blockHeight","type":"uint256"},{"internalType":"bytes","name":"PoD","type":"bytes"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[],"name":"removeDelegation","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"}]"#;
 
 pub type Qrc20AbiResult<T> = Result<T, MmError<Qrc20AbiError>>;
-
-lazy_static! {
-    pub static ref QTUM_DELEGATE_CONTRACT: Contract = Contract::load(QTUM_DELEGATE_CONTRACT_ABI.as_bytes()).unwrap();
-    pub static ref QTUM_DELEGATE_CONTRACT_ADDRESS: H160 =
-        ethabi::Address::from_str("0000000000000000000000000000000000000086").unwrap();
-}
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Qrc20ActivationParams {
@@ -1232,15 +1222,15 @@ pub fn contract_addr_into_rpc_format(address: &H160) -> H160Json { H160Json::fro
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Qrc20FeeDetails {
     /// Coin name
-    coin: String,
+    pub coin: String,
     /// Standard UTXO miner fee based on transaction size
-    miner_fee: BigDecimal,
+    pub miner_fee: BigDecimal,
     /// Gas limit in satoshi.
-    gas_limit: u64,
+    pub gas_limit: u64,
     /// Gas price in satoshi.
-    gas_price: u64,
+    pub gas_price: u64,
     /// Total used gas.
-    total_gas_fee: BigDecimal,
+    pub total_gas_fee: BigDecimal,
 }
 
 async fn qrc20_withdraw(coin: Qrc20Coin, req: WithdrawRequest) -> WithdrawResult {
@@ -1396,83 +1386,5 @@ fn transfer_event_from_log(log: &LogEntry) -> Result<TransferEventDetails, Strin
         amount,
         sender,
         receiver,
-    })
-}
-
-// Qtum Delegate Transaction - lock UTXO mutex before calling this function
-pub async fn generate_delegate_qrc20_delegation_transaction_from_qtum(
-    coin: &QtumCoin,
-    contract_outputs: Vec<ContractCallOutput>,
-    to_address: String,
-    gas_limit: u64,
-    transaction_type: TransactionType,
-) -> DelegationResult {
-    let utxo = coin.as_ref();
-    let (unspents, _) = coin.ordered_mature_unspents(&utxo.my_address).await?;
-    let mut gas_fee = 0;
-    let mut outputs = Vec::with_capacity(contract_outputs.len());
-    for output in contract_outputs {
-        gas_fee += output.gas_limit * output.gas_price;
-        outputs.push(TransactionOutput::from(output));
-    }
-
-    let (unsigned, data) = UtxoTxBuilder::new(coin)
-        .add_available_inputs(unspents)
-        .add_outputs(outputs)
-        .with_gas_fee(gas_fee)
-        .with_dust(QRC20_DUST)
-        .build()
-        .await
-        .mm_err(|gen_tx_error| {
-            DelegationError::from_generate_tx_error(gen_tx_error, coin.ticker().to_string(), utxo.decimals)
-        })?;
-
-    let prev_script = ScriptBuilder::build_p2pkh(&utxo.my_address.hash);
-    let signed = sign_tx(
-        unsigned,
-        &utxo.key_pair,
-        prev_script,
-        utxo.conf.signature_version,
-        utxo.conf.fork_id,
-    )
-    .map_to_mm(DelegationError::InternalError)?;
-
-    let miner_fee = data.fee_amount + data.unused_change.unwrap_or_default();
-    let generated_tx = GenerateQrc20TxResult {
-        signed,
-        miner_fee,
-        gas_fee,
-    };
-    let fee_details = Qrc20FeeDetails {
-        // QRC20 fees are paid in base platform currency (in particular Qtum)
-        coin: coin.ticker().to_string(),
-        miner_fee: utxo_common::big_decimal_from_sat(generated_tx.miner_fee as i64, utxo.decimals),
-        gas_limit,
-        gas_price: QRC20_GAS_PRICE_DEFAULT,
-        total_gas_fee: utxo_common::big_decimal_from_sat(generated_tx.gas_fee as i64, utxo.decimals),
-    };
-    let my_address = coin.my_address().map_to_mm(DelegationError::InternalError)?;
-
-    // Delegation transaction always transfer all the UTXO's to ourselves
-    let qtum_amount = coin.my_spendable_balance().compat().await?;
-    let received_by_me = qtum_amount.clone();
-    let my_balance_change = &received_by_me - &qtum_amount;
-
-    Ok(TransactionDetails {
-        tx_hex: serialize(&generated_tx.signed).into(),
-        tx_hash: generated_tx.signed.hash().reversed().to_vec().into(),
-        from: vec![my_address],
-        to: vec![to_address],
-        total_amount: qtum_amount.clone(),
-        spent_by_me: qtum_amount,
-        received_by_me,
-        my_balance_change,
-        block_height: 0,
-        timestamp: now_ms() / 1000,
-        fee_details: Some(fee_details.into()),
-        coin: coin.ticker().to_string(),
-        internal_id: vec![].into(),
-        kmd_rewards: None,
-        transaction_type,
     })
 }
